% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as_ellmer_tool}
\alias{as_ellmer_tool}
\title{Convert a deep tool to an ellmer tool}
\usage{
as_ellmer_tool(tool)
}
\arguments{
\item{tool}{A \code{deep_tool}.}
}
\value{
An \code{ellmer} tool definition.
}
\description{
Convert a deep tool to an ellmer tool
}
