% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitl.R
\name{decision_edit}
\alias{decision_edit}
\title{Human edit decision}
\usage{
decision_edit(arguments)
}
\arguments{
\item{arguments}{Replacement tool arguments.}
}
\value{
A decision object.
}
\description{
Human edit decision
}
