% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitl.R
\name{decision_reject}
\alias{decision_reject}
\title{Human rejection decision}
\usage{
decision_reject(message = "Rejected by user.")
}
\arguments{
\item{message}{Message returned to the model.}
}
\value{
A decision object.
}
\description{
Human rejection decision
}
