% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitl.R
\name{decision_respond}
\alias{decision_respond}
\title{Human response decision}
\usage{
decision_respond(message)
}
\arguments{
\item{message}{Message returned to the model.}
}
\value{
A decision object.
}
\description{
Human response decision
}
