% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{fs_permission}
\alias{fs_permission}
\title{Create a filesystem permission rule}
\usage{
fs_permission(operations, paths, mode = c("allow", "deny", "interrupt"))
}
\arguments{
\item{operations}{Character vector of operations such as \code{"read"} or \code{"write"}.}

\item{paths}{Character vector of virtual path glob patterns.}

\item{mode}{One of \code{"allow"}, \code{"deny"}, or \code{"interrupt"}.}
}
\value{
A permission rule.
}
\description{
Create a filesystem permission rule
}
