% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{mcp_tools}
\alias{mcp_tools}
\title{Return optional MCP-backed tools}
\usage{
mcp_tools(config = NULL, tools = list(), ...)
}
\arguments{
\item{config}{Optional mcptools MCP server configuration path.}

\item{tools}{Optional list of already materialized \code{\link[=deep_tool]{deep_tool()}} objects.}

\item{...}{Reserved for future \code{{mcptools}} server configuration.}
}
\value{
A list of tools.
}
\description{
Return optional MCP-backed tools
}
