% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag.R
\name{rag_tools}
\alias{rag_tools}
\title{Return optional RAG tools}
\usage{
rag_tools(search_fun = NULL, store = NULL, top_k = 3L, ...)
}
\arguments{
\item{search_fun}{Optional search function to wrap with \code{\link[=rag_tool]{rag_tool()}}.}

\item{store}{Optional Ragnar store object or path.}

\item{top_k}{Default number of chunks to retrieve from a Ragnar store.}

\item{...}{Additional arguments passed to \code{\link[=rag_tool]{rag_tool()}}.}
}
\value{
A list of tools.
}
\description{
Return optional RAG tools
}
