% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{rds_checkpointer}
\alias{rds_checkpointer}
\title{Create an RDS checkpointer}
\usage{
rds_checkpointer(path)
}
\arguments{
\item{path}{Destination \code{.rds} file for thread state.}
}
\value{
A checkpointer object.
}
\description{
Create an RDS checkpointer
}
