% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-rds-store.R
\name{rds_store_backend}
\alias{rds_store_backend}
\title{Create a durable RDS-backed virtual filesystem backend}
\usage{
rds_store_backend(path)
}
\arguments{
\item{path}{RDS file path.}
}
\value{
A backend object.
}
\description{
Create a durable RDS-backed virtual filesystem backend
}
